+++ src/main.cpp
@@ -22,8 +22,8 @@ int main(int argc, char ** argv)
     SDL_SetTextureBlendMode(texture, SDL_BLENDMODE_BLEND);
     // SDL_Texture * path = SDL_CreateTexture(renderer,
     // SDL_PIXELFORMAT_ARGB8888, SDL_TEXTUREACCESS_STATIC, WIDTH, HEIGHT);
-    std::pair<float, float> centers[10];
-    std::pair<float, float> speeds[10];
+    std::pair<double, double> centers[10];
+    std::pair<double, double> speeds[10];
     Ball* A[10];
     for(int i = 0; i < 10; ++i) {
     A[i] = new Ball(renderer, (rand() % (201)), (rand() % (201)), i); // Initialise each Ball object
